requirements.txt
yowsup2>=2.5.0